+++ src/api/consumer/lib/init.js
@@ -1,7 +1,6 @@
 /** @flow */
 import fs from 'fs-extra';
 import { Consumer } from '../../../consumer';
-import BitMap from '../../../consumer/bit-map/bit-map';
 import { Repository } from '../../../scope/objects';
 import { isDirEmpty } from '../../../utils';
 import ObjectsWithoutConsumer from './exceptions/objects-without-consumer';
@@ -17,25 +16,25 @@ export default (async function init(
   if (reset || resetHard) {
     await Consumer.reset(absPath, resetHard, noGit);
     overrideBitJson = true;
-  } else if (!force) {
-    await throwForOutOfSyncScope(absPath, noGit);
   }
   const consumer: Consumer = await Consumer.create(absPath, noGit);
+  if (!force) {
+    await throwForOutOfSyncScope(consumer);
+  }
   return consumer.write({ overrideBitJson });
 });
 
 /**
- * throw an error when .bitmap is deleted but a scope has objects.
+ * throw an error when .bitmap is empty but a scope has objects.
  * a user may got into this state for reasons such as:
  * 1. deleting manually .bitmap hoping to re-start Bit from scratch. (probably unaware of `--reset-hard` flag).
  * 2. switching to a branch where Bit wasn't initialized
  * in which case, it's better to stop and show an error describing what needs to be done.
  * it can always be ignored by entering `--force` flag.
  */
-async function throwForOutOfSyncScope(absPath: string, noGit: boolean): Promise<void> {
-  const { currentLocation: bitMapLocation } = BitMap.getBitMapLocation(absPath);
-  if (bitMapLocation) return;
-  const scopePath = Consumer._getScopePath(absPath, noGit);
+async function throwForOutOfSyncScope(consumer: Consumer): Promise<void> {
+  if (!consumer.bitMap.isEmpty()) return;
+  const scopePath = consumer.scope.getPath();
   const objectsPath = Repository.getPathByScopePath(scopePath);
   const dirExist = await fs.pathExists(objectsPath);
   if (!dirExist) return;