+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - generate package.json with the main property on node_modules for authored exported components instead of creating an entry point file
+- obsolete the "dependencies" property of consumer bit.json
+- avoid installing "undefined" npm package when importing authored components
+
+## [14.0.4-dev.1] - 2019-03-14
+
 - add authentication fallback to ssh-key in case the ssh-agent is enabled but failed to authenticate
 - improve authentication error message to clearly indicate the various strategies failures